Add kuadrant dependencies to OLM installation
Adds the generated kuadrant-controller installation manifests (CRD,RBAC
roles etc..) to the kuadrant-operator OLM bundle. This makes the controller
installation part of the OLM installation of the operator itself.
Authorino and Limitador are added as OLM dependencies to the operator
and defined in the bundles metadata/dependencies.yaml. The defined versions of
these componenets must be available in a catalog source avaiolable to
the target namesapce in order for the kuadrant-operator to install.

Makefile updates:

* Set Version to 0.0.0
* Set IMAGE_TAG to latest for 0.0.0
* Set REGISTRY(quay.io) and ORG(kuadrant) and use to build up IMAGE_TAG_BASE
* Add authorino, limitador and controller dependencies kustomize config
  for development installations.
* Add script to generate bundle dependencies file bundle/metadata/dependencies.yaml.
* Add make target `bundle-dependencies` to update dependencies.yaml with correct versions form Makefile.
* Update BUNDLE_IMGS to include limitadir and authorino bundles