fix: isAuthPolicyEnforcedCondition failing in eventually (#629)
The test "Route and Gateway AuthPolicies exist. Gateway AuthPolicy updated to remove overrides." is flaky since isAuthPolicyEnforcedCondition preveiously was not a function that can be executed continuously in an Eventually block so sometimes the AuthPolicy may not be enforced by the time the condition is checked. Since previously it was only checked once, it will always timeout even if the AuthPolicy eventually transitions to the expected status condition. Changing the function to return func() bool should fix this.