FEAT: create configmap of topology
This is a somewhat mess way to create the configmap.
A cleaner solution should be found which most likely will mean changing the DynamicClient.
For now this ia good enough.

This helps with running the container on a cluster as it is not trying to write topology.dot to disk.

Signed-off-by: Jim Fitzpatrick <jfitzpat@redhat.com>