fix: observability setup guide
Add examples/dashboards to the list of required kustomizations that must
be included to have grafana function.

Currently, if you do not apply the example dashboards kustomization the
grafana deployment will never start as there are configmaps missing that
are being mounted by the deployment causing the pod to never initialize
with various errors like `MountVolume.SetUp failed for volume
"grafana-business-user" : configmap "grafana-business-user" not found`.

Note: I do not believe this to be the correct long term fix, if grafana
requires those configmaps it should be part of the grafana kustomization
to create them.

Signed-off-by: Michael Nairn <mnairn@redhat.com>