ADD: Watch limitador CR on cluster
Adds a watch for the limitador CR on the cluster. This watches all
limitador CRDs across the cluster. Later this will be filtered down to be
more focused based of labels or owner reference. Owner reference
filtering is still to be explored.

Linking function adds the only expected limitador CR into the topology
based on the hard-coded name for that CR.

Signed-off-by: Jim Fitzpatrick <jfitzpat@redhat.com>