tests: Add --race option to ginkgo run
This will cause the test suite to fail if any race condition is detected
during the execution of the tests.

Note: When the tests are run in parallel this does not output a
descriptive message for the failure. If it fails for seemingly random
reasons you should try running the test suite without parallel processes
enabled.

Signed-off-by: Michael Nairn <mnairn@redhat.com>