Add canUpdateDNSRecord function
Returns true if an existing record can knowingly be updated to a desired
state based on the differences of the specs.

Current known reasons for not being able to update are:

* RootHost updates
* Endpoint record type changes (A -> CNAME etc..)

In both these cases, and any others that may be added to
`canUpdateDNSRecord` the current record should be deleted before doing a
create of the desired record.

Signed-off-by: Michael Nairn <mnairn@redhat.com>