fix: Remove dnspolicy lb immutable validation
Removes the validation on the loadBalancing field of the dnspolicy spec.
Updates made as part of section name support allow the controller to
handle changes to endpoint record types. It is now possible to switch
from what we used to refer to as simple policy types (loadBalancing
field not set) to loadbalanced (loadBalancing field set) and vice versa
without deleting and re-creating the policy.

Signed-off-by: Michael Nairn <mnairn@redhat.com>