test: dnspolicy scale testing (installation)
Adds a new installation overlay `namespaced-dns-operator` that when
applied to a cluster will install multiple instances of the dns operator
in different namepsaces. Each instance is configured to only watch for
dns resources (DNSRecords etc..) in that single namespace.

Templates the kuadrant and dns operator versions in config/install
overlays. These should be kept in sync with the branch/tag they are for
to allow more easily installing a particular version or latest.
These will be automicatically updated as part of the release creation.

Signed-off-by: Michael Nairn <mnairn@redhat.com>