FIX: sort.Sort not sorting correctly
The `sort.Sort` was failing to sort the list of limitador limits in a
standard fashion. This was causing multiple updates to the limitador CR.
By using `slices.SortFunc` a custom compare function is used, and now the
sorting happens as expected.

Extra logging has been added to help future investigations located
issues in the limits reconcile.

Signed-off-by: Jim Fitzpatrick <jfitzpat@redhat.com>