FIX: Status update error
If there are numerous resources removed from the cluster at the
same time we can sometimes be trying to update resources that have
already being removed. This is not an error, just noisy. An info log has
being added to cover when this happens.

By rights, I think this should be a warning message, but that is not a log level that we use.

Signed-off-by: Jim Fitzpatrick <jfitzpat@redhat.com>