dev: Allow setting metallb CIDR value for local env
The default CIDR value (/28) allows for 16 ips to be assigned which is
fine for the majority of dev/test/demo needs. However, when trying to
test at scale it is useful to be able to create more Gateways on a
single cluster and therefore more IPs can be required.

Changes here allow the CIDR and number of IP values used by the local
setup script to be set as required, default is the same as before (/28 &
16 IPs).

```
make local-setup SUBNET_OFFSET=0 CIDR=26 NUM_IPS=64
```

Signed-off-by: Michael Nairn <mnairn@redhat.com>