ADD: counterStorage UnSet
When the `counterStorage` is removed from the kuadrant resource it is also
removed from the limitador resource. If the user manually adds the
storage configuration to the limitador resource the change will not be
reverted if the feature is disabled in the kuadrant resource.

**EDGE CASE**
If the operator is scaled down when the user removes the
`counterStorage` from the kuadrant resource, and then restarts the
operator. The operator will not clean up the resources, it assumes it
was existing user configuration.

Signed-off-by: Jim Fitzpatrick <jfitzpat@redhat.com>