FIX: integration test update
If there is an error writing the resource back to the cluster the
status for the resilience should not be updated to ensure that the next
iteration can correctly write back to the cluster.

FIX: logic on cleanup

The cleanup check would fail if the resource had failed to write to the
cluster on the first instances.

Signed-off-by: Jim Fitzpatrick <jfitzpat@redhat.com>