FIX: status going to user defined
If the number of replicas is removed from the limitador CR, the
limitador will not scale the number of replicas to the default of one.

We need to update the limitador CR twice, and in order to do this a
sleep was added.

There could be a better way to do this over adding a sleep, but I
couldn't find a reliable way.

Signed-off-by: Jim Fitzpatrick <jfitzpat@redhat.com>